Refactor JoinGroupPage from GroupRankingPage
